Solvint for(int i issues

+++ g0/build.cpp
@@ -54,7 +54,7 @@ void setTestLocation(std::string newTL)
     //invert
     invertedLocation.clear();
 
-    for (int i = 0; i < newTL.length(); ++i)
+    for (size_t i = 0; i < newTL.length(); ++i)
         if (newTL[i]=='/')
             invertedLocation.push_back('\\');
         else
